+++ common/src/main/java/com/samsthenerd/hexgloop/casting/wehavelociathome/IContextHelper.java
@@ -1,0 +1,11 @@
+package com.samsthenerd.hexgloop.casting.wehavelociathome;
+
+import java.util.List;
+
+import net.minecraft.util.math.BlockPos;
+
+public interface IContextHelper {
+    public List<BlockPos> getChestRefs();
+
+    public void addChestRef(BlockPos pos);
+}