+++ common/src/main/java/com/samsthenerd/hexgloop/network/HexGloopNetwork.java
@@ -3,11 +3,14 @@ package com.samsthenerd.hexgloop.network;
 import java.util.List;
 
 import com.samsthenerd.hexgloop.HexGloop;
+import com.samsthenerd.hexgloop.mixins.orchard.MixinServerPlayerOrchard;
 import com.samsthenerd.hexgloop.network.booktweaks.BookScrollHandlers;
 import com.samsthenerd.hexgloop.network.booktweaks.BookScrollHandlersClient;
 
 import at.petrak.hexcasting.api.spell.casting.CastingHarness;
 import at.petrak.hexcasting.api.spell.casting.ResolvedPattern;
+import at.petrak.hexcasting.api.spell.math.HexDir;
+import at.petrak.hexcasting.api.spell.math.HexPattern;
 import at.petrak.hexcasting.common.network.MsgOpenSpellGuiAck;
 import at.petrak.hexcasting.xplat.IXplatAbstractions;
 import dev.architectury.networking.NetworkManager;
@@ -22,6 +25,8 @@ public class HexGloopNetwork {
     public static final Identifier CHANGE_WALL_SCROLL_ID = new Identifier(HexGloop.MOD_ID, "change_wall_scroll");
     public static final Identifier CLOSE_HEX_BOOK_ID = new Identifier(HexGloop.MOD_ID, "close_hex_book");
     public static final Identifier PROMPT_REPLACE_SCROLL_ID = new Identifier(HexGloop.MOD_ID, "prompt_replace_scroll");
+
+    public static final Identifier C2S_UPDATE_ORCHARD_ID = new Identifier(HexGloop.MOD_ID, "ctos_update_orchard"); // the important one
 
     public static void register(){
         // clears the grid - maybe not actually super useful
@@ -45,6 +50,14 @@ public class HexGloopNetwork {
             IXplatAbstractions.INSTANCE.sendPacketToPlayer(serverPlayer,
                 new MsgOpenSpellGuiAck(Hand.MAIN_HAND, patterns, descs.getFirst(), descs.getSecond(), descs.getThird(),
                     harness.getParenCount()));
+        });
+
+        NetworkManager.registerReceiver(NetworkManager.Side.C2S, C2S_UPDATE_ORCHARD_ID, (buf, context) -> {
+            PlayerEntity player = context.getPlayer();
+            if(!(player instanceof ServerPlayerEntity sPlayer)) return;
+            String angleSig = buf.readString();
+            boolean activated = buf.readBoolean();
+            ((MixinServerPlayerOrchard)(Object)sPlayer).setAssociation(HexPattern.fromAngles(angleSig, HexDir.EAST), activated);
         });
 
         NetworkManager.registerReceiver(NetworkManager.Side.C2S, CHANGE_WALL_SCROLL_ID, BookScrollHandlers::handleReplaceScroll);